Refactored ChessData\Cli\Seed\Users
